recast: store compact heightfield areas one byte per span
chf.areas held an int per span where the C++ original holds an unsigned
char. Area ids are small - RC_WALKABLE_AREA is 63 - so three of every
four bytes were padding being dragged through the cache.

It matters because of who reads it. ExpandRegions touches areas once per
span per direction, which makes it the heaviest consumer of the array in
the build, and the region stages are the largest part of a tile bake.

Measured on a six-tile World of Warcraft corpus, three interleaved A/B
pairs run in drift-cancelling order on an idle machine (Release, 4 cores):

  RC_TIMER_BUILD_REGIONS_EXPAND     170.4/133.8/150.3 -> 125.5/128.6/127.7  (-16%)
  RC_TIMER_BUILD_REGIONS_WATERSHED  245.8/207.2/230.6 -> 199.4/202.8/201.1  (-12%)
  RC_TIMER_BUILD_REGIONS            276.8/238.5/263.8 -> 231.3/234.2/234.2  (-10%)

No sample in the new set overlaps the old set on any of those three.
Total bake moves about -2%, which is inside the run-to-run spread.

Two stages read slightly slower - FILTER_BORDER and FILTER_LOW_OBSTACLES,
both around +4% - but neither touches this array, and both run before the
compact heightfield exists. That is code layout, not cause.

MEDIAN_AREA, which sweeps areas in a 3x3 stencil, did not move. The win
is in the region build, not the filter.

Allocation is unchanged at 85 MB: the array was already pooled, so this
narrows the working set rather than the allocation rate. For the largest
tile in the corpus that is 16.0 MB of compact heightfield down to 14.3 MB,
against 7.4 MB for the same tile in C++.

Output is byte-identical: the SHA-256 of every produced tile is unchanged
across the corpus.

+++ src/DotRecast.Recast/RcAreas.cs
@@ -277,11 +277,11 @@ namespace DotRecast.Recast
 
             using var timer = context.ScopedTimer(RcTimerLabel.RC_TIMER_MEDIAN_AREA);
 
-            int[] areas = compactHeightfield.Rent<int>(compactHeightfield.spanCount);
+            byte[] areas = compactHeightfield.Rent<byte>(compactHeightfield.spanCount);
 
             // Hoisted: these are class fields reloaded on every innermost access.
             // compactHeightfield.areas is only replaced after the whole loop.
-            int[] srcAreas = compactHeightfield.areas;
+            byte[] srcAreas = compactHeightfield.areas;
             RcCompactCell[] cells = compactHeightfield.cells;
             RcCompactSpan[] spans = compactHeightfield.spans;
 
@@ -298,7 +298,7 @@ namespace DotRecast.Recast
                     for (int spanIndex = cell.index; spanIndex < maxSpanIndex; ++spanIndex)
                     {
                         ref RcCompactSpan span = ref spans[spanIndex];
-                        int centerArea = srcAreas[spanIndex];
+                        byte centerArea = srcAreas[spanIndex];
                         if (centerArea == RC_NULL_AREA)
                         {
                             areas[spanIndex] = centerArea;
@@ -355,7 +355,7 @@ namespace DotRecast.Recast
                         }
 
                         neighborAreas.InsertSort();
-                        areas[spanIndex] = neighborAreas[4];
+                        areas[spanIndex] = (byte)neighborAreas[4];
                     }
                 }
             }
@@ -454,7 +454,7 @@ namespace DotRecast.Recast
                         }
 
                         // Mark the span.
-                        compactHeightfield.areas[spanIndex] = areaId.Apply(compactHeightfield.areas[spanIndex]);
+                        compactHeightfield.areas[spanIndex] = (byte)areaId.Apply(compactHeightfield.areas[spanIndex]);
                     }
                 }
             }
@@ -578,7 +578,7 @@ namespace DotRecast.Recast
 
                         if (PointInPoly(verts, point))
                         {
-                            compactHeightfield.areas[spanIndex] = areaId.Apply(compactHeightfield.areas[spanIndex]);
+                            compactHeightfield.areas[spanIndex] = (byte)areaId.Apply(compactHeightfield.areas[spanIndex]);
                         }
                     }
                 }
@@ -703,7 +703,7 @@ namespace DotRecast.Recast
                         // Mark if y extents overlap.
                         if (span.y >= miny && span.y <= maxy)
                         {
-                            compactHeightfield.areas[spanIndex] = areaId.Apply(compactHeightfield.areas[spanIndex]);
+                            compactHeightfield.areas[spanIndex] = (byte)areaId.Apply(compactHeightfield.areas[spanIndex]);
                         }
                     }
                 }
